Remove unnecessary condition
As far as I know, we don't currently make any difference between an
undefined setting, or a defined setting with `nil` value, so it would be
meaningless to pass `nil` as a default here.

Even if this was useful, this is currently dead code, so it's better to
remove it until we find use case setting for it, in my opinion.